Code refactoring, moved attributes to root
