+++ Microsoft.Toolkit/Attributes/DoesNotReturnAttribute.cs
