+++ Microsoft.Toolkit/Attributes/DoesNotReturnIfAttribute.cs
